% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Cronbach's Coefficient Alpha}
\usage{
alpha(x)
}
\arguments{
\item{x}{A data frame or matrix containing data with rows as subjects and columns as items}
}
\value{
Cronbach's coefficient alpha and associated SEM
}
\description{
Compute the popular Cronbach's coefficient alpha and associated SEM
}
\details{
Alpha is one of several estimates of the internal consistency reliability of a test.
}
\examples{
data(data.u)
alpha(data.u)


}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. \emph{Psychometrika, 6,} 297-334.
}
