% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbach-alpha.R
\name{cronbach_alpha}
\alias{cronbach_alpha}
\title{Cronbach's Alpha}
\usage{
cronbach_alpha(dat)
}
\arguments{
\item{dat}{a data frame or matrix containing raw data}
}
\value{
Cronbach's Alpha reliability
}
\description{
A function to calculate Cronbach's Alpha reliability in CTT
}
\examples{


}
