% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CronbachAlpha.R
\name{CronbachAlpha}
\alias{CronbachAlpha}
\title{Cronbach's Alpha}
\usage{
CronbachAlpha(dat)
}
\arguments{
\item{dat}{a data frame or matrix containing raw data}
}
\value{
a reliability number
}
\description{
A function to calculate Cronbach's Alpha reliability in CTT
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
