% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem.binomial.R
\name{csem.binomial}
\alias{csem.binomial}
\title{CSEM and CSSEM with Binomial Model}
\usage{
csem.binomial(ni, ct = NULL)
}
\arguments{
\item{ni}{a numeric value indicating the number of items}

\item{ct}{A data frame or matrix containing conversion table with the first column
as raw scores and the second column as scale scores}
}
\value{
a list containing raw scores and corresponding CSEM using bimonial model,
and CSSEM if the conversion table is provided
}
\description{
Compute the CSEM and CSSEM using binomial model
}
\details{

}
\examples{
csem.binomial(40)
csem.binomial(40, ct.u)

}
\references{

}
