% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem.compound.binomial.R
\name{csem.compound.binomial}
\alias{csem.compound.binomial}
\title{CSEM with Compound Binomial Model}
\usage{
csem.compound.binomial(x, s)
}
\arguments{
\item{x}{A data frame or matrix containing data with rows as subjects and items as columns}

\item{s}{A vector containing number of items for each strat}
}
\value{
a list containing raw scores and corresponding CSEM using compound bimonial model
}
\description{
Compute the CSEM using compound binomial model
}
\details{

}
\examples{
data(data.m)
csem.compound.binomial(data.m, c(13, 12, 6))

}
\references{

}
