% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEMIRT.R
\name{CSEMIRT}
\alias{CSEMIRT}
\title{CSEM of IRT 2PL: EAP&MLE}
\usage{
CSEMIRT(theta, itemPara, estType)
}
\arguments{
\item{theta}{a vector, matrix, or data frame containing theta values}

\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}

\item{estType}{estimation method, MLE or EAP}
}
\value{
a data frame containing IRT EAP&MLE CSEM
}
\description{
A function to calculate CSEM of 2PL IRT with EAP&MLE
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
