% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEMLord.R
\name{CSEMLord}
\alias{CSEMLord}
\title{CSEM Lord Mehtod}
\usage{
CSEMLord(numOfItem)
}
\arguments{
\item{numOfItem}{a numeric number indicating number of items}
}
\value{
a data frame containing Lord CSEM
}
\description{
A function to calculate Conditional Standard Error of Measurement in CTT
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
