% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSEMIRTPoly.R
\name{CSSEMIRTPoly}
\alias{CSSEMIRTPoly}
\title{CSSEM IRT Polynomial Method}
\usage{
CSSEMIRTPoly(itemPara, convTable, K, estType)
}
\arguments{
\item{convTable}{a data frame or matrix containing conversion table of raw score to scale score}

\item{K}{a numeric number indicating degree of polynomial regression}

\item{estType}{estimation method, MLE or EAP}

\item{numOfItem}{a numeric number indicating number of items}
}
\value{
a list containing R square and CSSEM using Polynomial Method
}
\description{
A function to calculate CSEM for Scale Scores in IRT using Polynomial Method
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
