% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSEMKolen.R
\name{CSSEMKolen}
\alias{CSSEMKolen}
\title{CSSEM Kolen's Method}
\usage{
CSSEMKolen(itemPara, convTable)
}
\arguments{
\item{convTable}{a data frame or matrix containing conversion table of raw score to scale score}

\item{numOfItem}{a numeric number indicating number of items}
}
\value{
a data frame containing CSSEM using Kolen's Method
}
\description{
A function to calculate CSEM for Scale Scores in IRT using Kolen's method
True scale score
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
