% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSEMPolynomial.R
\name{CSSEMPolynomial}
\alias{CSSEMPolynomial}
\title{CSSEM Polynomial Method}
\usage{
CSSEMPolynomial(numOfItem, convTable, K)
}
\arguments{
\item{numOfItem}{a numeric number indicating number of items}

\item{convTable}{a data frame or matrix containing conversion table of raw score to scale score}

\item{K}{a numeric number indicating highest degree of polynomial regression}
}
\value{
a data frame containing CSSEM using Polynomial Method
}
\description{
A function to calculate CSEM for Scale Scores using Polynomial Method based on Lord's CSEM
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
