% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct.u}
\alias{ct.u}
\title{Conversion table for unidimensional data.}
\format{
A data frame with 41 rows and 2 variables:
\describe{
\item{x}{raw score}
\item{ss}{scale score}
}
}
\usage{
ct.u
}
\description{
A dataset containing the conversion table for the unidimensional data, with first column
as raw scores and second column as scale scores
}
\keyword{datasets}
