% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.m}
\alias{data.m}
\title{Multidimensional data.}
\format{
A data frame with 3000 rows and 31 variables:
\describe{
\item{V1}{responses to item 1}
\item{V2}{responses to item 2}\cr
...\cr
\item{V31}{responses to item 31}
}
}
\usage{
data.m
}
\description{
A dataset containing the responses of 3000 subjects to 31 items on three subscales (13, 12, 5 items respectively).
}
\keyword{datasets}
