% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.u}
\alias{data.u}
\title{Unidimensional data.}
\format{
A data frame with 3000 rows and 40 variables:
\describe{
\item{V1}{responses to item 1}
\item{V2}{responses to item 2}\cr
...\cr
\item{V40}{responses to item 40}
}
}
\usage{
data.u
}
\description{
A dataset containing the responses of 3000 subjects to 40 items
}
\keyword{datasets}
