% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feldt.R
\name{feldt}
\alias{feldt}
\title{Feldt's Coefficient}
\usage{
feldt(x)
}
\arguments{
\item{x}{A data frame or matrix containing data with rows as subjects and columns as items}
}
\value{
Feldt's coefficient
}
\description{
Compute the Feldt's coefficient
}
\details{
Feldt's coefficient is one of several estimates of the internal consistency reliability of a test.
}
\examples{
data(data.u)
feldt(data.u)


}
\references{
Brennan, R. L.(Ed.). (2006). \emph{Educational Measurement} (4th ed.). Westport, CT:Praeger Publishers.
}
