% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Information for IRT model}
\usage{
info(theta, ip, est)
}
\arguments{
\item{theta}{Ability vector}

\item{ip}{a dataframe or matrix with item parameters of sequence b, a, and c,
a is on the 1.702 metric}

\item{est}{estimation method, MLE or EAP}
}
\value{
a data frame containing info
}
\description{
A function to calculate information for IRT model
}
