% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KolenRelIRT.R
\name{KolenRelIRT}
\alias{KolenRelIRT}
\title{IRT Reliability of Kolen's Method}
\usage{
KolenRelIRT(itemPara, convTable)
}
\arguments{
\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}

\item{convTable}{matrix or data frame contains raw score to scale score}
}
\value{
a reliability number
}
\description{
A function to calculate IRT Reliability of Kolen's Method
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
