% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LordWingersky.R
\name{LordWingersky}
\alias{LordWingersky}
\title{Lord Wingersky recursive formula}
\usage{
LordWingersky(probs)
}
\arguments{
\item{probs}{probabilities that a theta value can correctly answer each item}
}
\value{
probability of raw score distribution for the specfic theta
}
\description{
A function to calculate raw score distribution conditional on quadrature points
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
