% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalRelIRT.R
\name{MarginalRelIRT}
\alias{MarginalRelIRT}
\title{Marginal Reliability of IRT 2PL: MLE&EAP}
\usage{
MarginalRelIRT(itemPara, estType)
}
\arguments{
\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}

\item{estType}{estimation method, MLE or EAP}
}
\value{
a reliability number
}
\description{
A function to calculate marginal reliability of 2PL IRT with MLE&EAP
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
