% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalRelMLE.R
\name{MarginalRelMLE}
\alias{MarginalRelMLE}
\title{Marginal Reliability of IRT 2PL: MLE}
\usage{
MarginalRelMLE(itemPara)
}
\arguments{
\item{itemPara}{a text file}
}
\value{
a reliability number
}
\description{
A function to calculate marginal reliability of 2PL IRT with MLE
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
