% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-quadra.R
\name{normal_quadra}
\alias{normal_quadra}
\title{Gaussian quadrature points and weights}
\usage{
normal_quadra(n, mm)
}
\arguments{
\item{n}{Number of quadrature points}

\item{mm}{Absolute value of minumum or maximum of quadrature points}
}
\value{
A list of Gaussian quadrature points and corresponding weights
}
\description{
Generate Gaussian quadrature points and corresponding weights
}
\examples{
normal_quadra(41, 5)


}
