% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalQuadraPoints.R
\name{NormalQuadraPoints}
\alias{NormalQuadraPoints}
\title{normal quadrature points and normalized weights}
\usage{
NormalQuadraPoints(n)
}
\arguments{
\item{n}{number of quadrature points evenly distributed from -5 to +5}
}
\value{
a list containing nodes and normalized weights
}
\description{
A function to generate quadrature points and normalized weights
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
