% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomialMethod.R
\name{PolynomialMethod}
\alias{PolynomialMethod}
\title{Polynomial Method for CSSEM}
\usage{
PolynomialMethod(cssemDat, K)
}
\arguments{
\item{cssemDat}{a ata frame or matrix containing conversion table of raw score to scale score, and csem of raw score}

\item{K}{degree of polynomial regression}
}
\value{
a data frame containing CSSEM using Polynomial Method using different k values
}
\description{
A function to implement polynomial method in calculating CSSEM
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
