% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelIRTPoly.R
\name{RelIRTPoly}
\alias{RelIRTPoly}
\title{IRT Reliability Polynomial Method}
\usage{
RelIRTPoly(itemPara, convTable, K, estType, rawData)
}
\arguments{
\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}

\item{convTable}{a data frame or matrix containing conversion table of raw score to scale score}

\item{K}{a numeric number indicating highest degree of polynomial regression}

\item{estType}{estimation method, MLE or EAP}

\item{rawData}{raw data matrix}
}
\value{
a data frame containing CSSEM using Polynomial Method
}
\description{
A function to calculate reliability for Scale Scores using Polynomial Method based on IRT CSEM
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
