% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman.brown.R
\name{spearman.brown}
\alias{spearman.brown}
\title{Spearman-Brown Prophecy Formula}
\usage{
spearman.brown(rxx, input, type = "r")
}
\arguments{
\item{rxx}{a numeric value indicating the original reliability}

\item{input}{a numeric value indicating the desired reliability or a ratio of
new test length to original test length}

\item{type}{if type = "r", the new test reliability will be computed; if type = "l",
the function will return a ratio of the new test length to original test length}
}
\value{
the desired reliability or the desired reliability or a ratio of
new test length to original test length, depending on the value of argument type
}
\description{
Compute the new reliability when test length changes given the original test reliability,
or the ratio of new test length to original test length when specific reliability is desired.
}
\details{

}
\examples{
spearman.brown(0.8, 2.25, "r")
spearman.brown(0.8, 0.90, "l")

}
\references{

}
