% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.alpha.R
\name{stratified.alpha}
\alias{stratified.alpha}
\title{Stratified Cronbach's Coefficient Alpha}
\usage{
stratified.alpha(x, s)
}
\arguments{
\item{x}{A data frame or matrix containing data with rows as subjects and items as columns}

\item{s}{A vector containing number of items for each strat}
}
\value{
stratified Cronbach's coefficient alpha
}
\description{
Compute the stratified Cronbach's coefficient alpha
}
\details{
Stratified alpha is one of several estimates of the internal consistency reliability of a test.
}
\examples{
data(data.m)
stratified.alpha(data.m, c(13, 12, 6))


}
\references{
Cronbach, L. J., Schoenemann, P., & McKie, D. (1965). Alpha coefficient for stratified-parallel
tests. \emph{Educational and Psychological Measurement, 25,} 291-312.
}
