% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.feldt.R
\name{stratified.feldt}
\alias{stratified.feldt}
\title{Stratified Feldt's Coefficient}
\usage{
stratified.feldt(x, s)
}
\arguments{
\item{x}{A data frame or matrix containing data with rows as subjects and items as columns}

\item{s}{A vector containing number of items for each strat}
}
\value{
stratified Feldt's coefficient
}
\description{
Compute the stratified Feldt's coefficient
}
\details{
Stratified Feldt's coefficient is one of several estimates of the internal consistency reliability of a test.
}
\examples{
data(data.m)
stratified.feldt(data.m, c(13, 12, 6))


}
\references{

}
