% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRelIRT.R
\name{TestRelIRT}
\alias{TestRelIRT}
\title{TestRelIRT}
\usage{
TestRelIRT(itemPara)
}
\arguments{
\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}
}
\value{
a reliability number
}
\description{
A function to calculate test reliability of IRT
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
