% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feldt.R
\name{Feldt}
\alias{Feldt}
\title{Feldt}
\usage{
Feldt(dat)
}
\arguments{
\item{dat}{a data frame or matrix containing raw data}
}
\value{
a reliability number
}
\description{
A function to calculate Feldt reliability in CTT
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
