% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info.R
\name{Info}
\alias{Info}
\title{IRT Information for 2PL model}
\usage{
Info(theta, itemPara, estType)
}
\arguments{
\item{theta}{ability theta vector}

\item{itemPara}{a text file with parameters of sequence b and a, a is on the 1.702 metric}

\item{estType}{estimation method, MLE or EAP}
}
\value{
a data frame containing info
}
\description{
A function to calculate information for 2PL model
}
\author{
{Huan Liu, University of Iowa, \email{huan-liu-1@uiowa.edu}}
}
